Go ファイルは機能分けにする
